** 게시물 삭제 다이얼로그 추가함  **
 - 간단. AlertDialog 추가해서 확인하고 삭제하도록
 - 앱 아이콘 변경
 - 앱 이름 변경
 - 페이스북, 구글 로그인 모두 가능.
 - 구글 플레이 콘솔에서 릴리즈에 사용한 서명 SHA-1키 확인 가능 ***

+++ app/src/main/java/com/chodingcoding/happygraudate/navigation/DetailViewFragment.kt
@@ -1,7 +1,9 @@
 package com.chodingcoding.happygraudate.navigation
 
+import android.app.AlertDialog
 import android.content.Intent
 import android.os.Bundle
+import android.view.ContextThemeWrapper
 import android.view.LayoutInflater
 
 import android.view.View
@@ -118,21 +120,37 @@ class DetailViewFragment : Fragment(){
                 viewholder.detailviewitem_setting_img.visibility = View.VISIBLE
                 viewholder.detailviewitem_setting_img.setOnClickListener {
 
+                    val builder = AlertDialog.Builder(ContextThemeWrapper(activity, R.style.Theme_AppCompat_Light_Dialog))
+                    builder.setTitle("게시물 삭제")
+                    builder.setMessage("정말로 삭제하시겠습니까?")
 
-                    var ref = firebaseStorage?.reference?.child("images")?.child(
-                        contentDTOs[position].fileName!!)
-                    ref?.delete()?.addOnCompleteListener {
-                        task ->
-                        if(task.isSuccessful){
-                            firestore?.collection("images")?.document(contentDTOs[position].fileName!!)?.delete()?.addOnCompleteListener {
+                    /*builder.setPositiveButton("확인") {dialog, id ->
+                    }
+                    builder.setNegativeButton("취소") {dialog, id ->
+                    }*/
+                    builder.setPositiveButton("확인") { _, _ ->
+                        var ref = firebaseStorage?.reference?.child("images")?.child(
+                            contentDTOs[position].fileName!!)
+                        ref?.delete()?.addOnCompleteListener {
                                 task ->
-                                if(task.isSuccessful){
-                                    Toast.makeText(activity, "게시물 삭제가 완료되었습니다.", Toast.LENGTH_LONG).show()
-                                }
+                            if(task.isSuccessful){
+                                firestore?.collection("images")?.document(contentDTOs[position].fileName!!)?.delete()?.addOnCompleteListener {
+                                        task ->
+                                    if(task.isSuccessful){
+                                        Toast.makeText(activity, "게시물 삭제가 완료되었습니다.", Toast.LENGTH_LONG).show()
+                                    }
 
+                                }
                             }
                         }
+
                     }
+                    builder.setNegativeButton("취소") { _, _ ->
+
+                    }
+
+                    builder.show()
+
